\name{symSEM-package}
\alias{symSEM-package}
\alias{symSEM}
\docType{package}
\title{
\packageTitle{symSEM}Symbolic Computation for Structural Equation Models
}
\description{
Some functions of symbolic computation for structural equation models (SEM).
}
\details{
\tabular{ll}{
Package: \tab metaSEM\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2020-07-02\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Maintainer: Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{ package }
