\name{symSEM-package}
\alias{symSEM-package}
\alias{symSEM}
\docType{package}
\title{
\packageTitle{symSEM}Symbolic Computation for Structural Equation Models
}
\description{
  A collection of functions for symbolic computation using 'Ryacas'
  package for structural equation models. This package includes
  functions to calculate model-implied covariance (and correlation)
  matrix and sampling covariance matrix of functions of variables using
  the first-order Taylor approximation. Reference: McArdle and McDonald (1984) <doi:10.1111/j.2044-8317.1984.tb00802.x>.
}
\details{
\tabular{ll}{
Package: \tab symSEM\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2020-09-26\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Maintainer: Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{ package }
