% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symsem-package.R
\name{symSEM-package}
\alias{symSEM-package}
\title{A collection of functions for symbolic computation using 'caracas'
package for structural equation models and other statistical analyses.
Among its features is the ability to calculate the model-implied
covariance (and correlation) matrix and the sampling covariance matrix of
variable functions using the delta method.}
\description{
A collection of functions for symbolic computation using 'caracas'
package for structural equation models and other statistical analyses.
Among its features is the ability to calculate the model-implied
covariance (and correlation) matrix and the sampling covariance matrix of
variable functions using the delta method.
}
\note{
As 'caracas' uses 'SymPy" in the backend. Reserved words in SymP,
such as"lambda" and "I" are converted to some random strings first.
These random strings are converted back to R.
}
